Change some types in ArMapComponents etc. (number of lines, number of points) from int to size_t. Other type fixes.

+++ include/Aria/ArMap.h
@@ -295,7 +295,7 @@ public:
 
   AREXPORT virtual ArPose getMinPose(const char *scanType = ARMAP_DEFAULT_SCAN_TYPE);
   AREXPORT virtual ArPose getMaxPose(const char *scanType = ARMAP_DEFAULT_SCAN_TYPE);
-  AREXPORT virtual int getNumPoints(const char *scanType = ARMAP_DEFAULT_SCAN_TYPE);
+  AREXPORT virtual size_t getNumPoints(const char *scanType = ARMAP_DEFAULT_SCAN_TYPE);
   AREXPORT virtual bool isSortedPoints(const char *scanType = ARMAP_DEFAULT_SCAN_TYPE) const;
 
   AREXPORT virtual void setPoints(const std::vector<ArPose> *points,
@@ -308,7 +308,7 @@ public:
 
   AREXPORT virtual ArPose getLineMinPose(const char *scanType = ARMAP_DEFAULT_SCAN_TYPE);
   AREXPORT virtual ArPose getLineMaxPose(const char *scanType = ARMAP_DEFAULT_SCAN_TYPE);
-  AREXPORT virtual int getNumLines(const char *scanType = ARMAP_DEFAULT_SCAN_TYPE);
+  AREXPORT virtual size_t getNumLines(const char *scanType = ARMAP_DEFAULT_SCAN_TYPE);
   AREXPORT virtual bool isSortedLines(const char *scanType = ARMAP_DEFAULT_SCAN_TYPE) const;
    
   AREXPORT virtual void setLines(const std::vector<ArLineSegment> *lines,
@@ -328,12 +328,12 @@ public:
                                const char *scanType = ARMAP_DEFAULT_SCAN_TYPE);
 
   AREXPORT virtual void writePointsToFunctor
-                              (ArFunctor2<int, std::vector<ArPose> *> *functor,
+                              (ArFunctor2<size_t, std::vector<ArPose> *> *functor,
                                const char *scanType = ARMAP_DEFAULT_SCAN_TYPE,
                                ArFunctor1<const char *> *keywordFunctor = NULL);
 
    AREXPORT virtual void writeLinesToFunctor
- 		                          (ArFunctor2<int, std::vector<ArLineSegment> *> *functor,
+ 		                          (ArFunctor2<size_t, std::vector<ArLineSegment> *> *functor,
                                const char *scanType = ARMAP_DEFAULT_SCAN_TYPE,
                                ArFunctor1<const char *> *keywordFunctor = NULL);
